+++ tests/unit/AppInfo/RoutesTest.php
@@ -64,6 +64,22 @@ class Test extends TestCase  {
 					'url' => '/saml/selectUserBackEnd',
 					'verb' => 'GET',
 				],
+				[
+					'name' => 'Settings#getSamlProviderSettings',
+					'url' => '/settings/providerSettings/{providerId}',
+					'verb' => 'GET',
+					'defaults' => [
+						'providerId' => '1'
+					]
+				],
+				[
+					'name' => 'Settings#deleteSamlProviderSettings',
+					'url' => '/settings/providerSettings/{providerId}',
+					'verb' => 'DELETE',
+					'defaults' => [
+						'providerId' => '1'
+					]
+				],
 			],
 		];
 		$this->assertSame($expected, $routes);